+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de s
 - `QasActions`: alterado componente para Composition API.
 - `QasNestedFields`: alterado prop `fieldsProps` para aceitar uma função de callback.
 - `QasNestedFields`: alterado slot `custom-fields` para se chamar `after-fields`. O slot `after-fields` é renderizado após os campos do formulário.
+- `QasNestedFields`: alterado tipo da propriedade `formColumns` para aceitar `array`, `string` ou `object`.
 - `QasActionsMenu`: alterado componente para Composition API.
 - `QasAppBar`: alterado componente para Composition API.